+++ R/provider-claude.R
@@ -171,15 +171,23 @@ method(base_request, ProviderAnthropic) <- function(provider) {
     req <- req_headers(req, `anthropic-beta` = provider@beta_headers)
   }
 
-  # <https://docs.anthropic.com/en/api/errors>
-  req <- req_error(req, body = function(resp) {
-    if (resp_content_type(resp) == "application/json") {
-      json <- resp_body_json(resp)
-      paste0(json$error$message, " [", json$error$type, "]")
-    }
-  })
+  req <- base_request_error(provider, req)
 
   req
+}
+
+method(base_request_error, ProviderAnthropic) <- function(provider, req) {
+  req_error(req, body = anthropic_error_body)
+}
+
+# <https://docs.anthropic.com/en/api/errors>
+anthropic_error_body <- function(resp) {
+  if (identical(resp_content_type(resp), "application/json")) {
+    json <- resp_body_json(resp)
+    if (!is.null(json$error)) {
+      paste0(json$error$message, " [", json$error$type, "]")
+    }
+  }
 }